blocking way to consume events from pending queue
