+++ server/src/test/java/org/apache/iotdb/db/pipe/core/collector/PipeRealtimeCollectTest.java
@@ -29,7 +29,7 @@ import org.apache.iotdb.db.pipe.config.PipeCollectorConstant;
 import org.apache.iotdb.db.pipe.core.collector.realtime.PipeRealtimeDataRegionCollector;
 import org.apache.iotdb.db.pipe.core.collector.realtime.PipeRealtimeDataRegionHybridCollector;
 import org.apache.iotdb.db.pipe.core.collector.realtime.listener.PipeInsertionDataNodeListener;
-import org.apache.iotdb.db.pipe.task.queue.ListenableUnblockingPendingQueue;
+import org.apache.iotdb.db.pipe.task.queue.ListenableUnboundedBlockingPendingQueue;
 import org.apache.iotdb.db.wal.utils.WALEntryHandler;
 import org.apache.iotdb.pipe.api.customizer.PipeParameters;
 import org.apache.iotdb.pipe.api.event.Event;
@@ -103,16 +103,16 @@ public class PipeRealtimeCollectTest {
 
     try (PipeRealtimeDataRegionHybridCollector collector1 =
             new PipeRealtimeDataRegionHybridCollector(
-                null, new ListenableUnblockingPendingQueue<>());
+                null, new ListenableUnboundedBlockingPendingQueue<>());
         PipeRealtimeDataRegionHybridCollector collector2 =
             new PipeRealtimeDataRegionHybridCollector(
-                null, new ListenableUnblockingPendingQueue<>());
+                null, new ListenableUnboundedBlockingPendingQueue<>());
         PipeRealtimeDataRegionHybridCollector collector3 =
             new PipeRealtimeDataRegionHybridCollector(
-                null, new ListenableUnblockingPendingQueue<>());
+                null, new ListenableUnboundedBlockingPendingQueue<>());
         PipeRealtimeDataRegionHybridCollector collector4 =
             new PipeRealtimeDataRegionHybridCollector(
-                null, new ListenableUnblockingPendingQueue<>())) {
+                null, new ListenableUnboundedBlockingPendingQueue<>())) {
 
       collector1.customize(
           new PipeParameters(
